+++ packages/react-vtable/src/tables/base-table.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/display-name */
 // import * as VTable from '@visactor/vtable';
-import { VTable } from '../vtable';
+// import { VTable } from '../vtable';
 import React, { useState, useEffect, useRef, useImperativeHandle, useCallback } from 'react';
 import type { ContainerProps } from '../containers/withContainer';
 import withContainer from '../containers/withContainer';
@@ -9,7 +9,7 @@ import RootTableContext from '../context/table';
 import { isEqual, isNil, pickWithout } from '@visactor/vutils';
 import { toArray } from '../util';
 import { REACT_PRIVATE_PROPS } from '../constants';
-import type { IMarkElement } from '../components';
+import type { IMarkElement } from '../table-components';
 import type {
   EventsProps
   // LegendEventProps,
@@ -22,17 +22,23 @@ import type {
   // TableLifeCycleEventProps
 } from '../eventsUtils';
 import { bindEventsToTable, TABLE_EVENTS_KEYS, TABLE_EVENTS } from '../eventsUtils';
-import { VTableReactAttributePlugin } from '../components/custom/vtable-react-attribute-plugin';
-import { reactEnvModule } from '../components/custom/vtable-browser-env-contribution';
-const { container, isBrowserEnv } = VTable.VRender;
+import { VTableReactAttributePlugin } from '../table-components/custom/vtable-react-attribute-plugin';
+import { reactEnvModule } from '../table-components/custom/vtable-browser-env-contribution';
+import { container, isBrowserEnv } from '@visactor/vtable/es/vrender';
+import type {
+  ListTable,
+  PivotTable,
+  PivotChart,
+  ListTableConstructorOptions,
+  PivotTableConstructorOptions,
+  PivotChartConstructorOptions
+} from '@visactor/vtable';
 
-export type IVTable = VTable.ListTable | VTable.PivotTable | VTable.PivotChart;
-export type IOption =
-  | VTable.ListTableConstructorOptions
-  | VTable.PivotTableConstructorOptions
-  | VTable.PivotChartConstructorOptions;
+export type IVTable = ListTable | PivotTable | PivotChart;
+export type IOption = ListTableConstructorOptions | PivotTableConstructorOptions | PivotChartConstructorOptions;
 
 export interface BaseTableProps extends EventsProps {
+  vtableConstrouctor?: any;
   type?: string;
   /** 上层container */
   container?: HTMLDivElement;
@@ -69,7 +75,8 @@ const notOptionKeys = [
   'onReady',
   'option',
   'records',
-  'container'
+  'container',
+  'vtableConstrouctor'
 ];
 
 const getComponentId = (child: React.ReactNode, index: number) => {
@@ -78,7 +85,7 @@ const getComponentId = (child: React.ReactNode, index: number) => {
 };
 
 const parseOptionFromChildren = (props: Props) => {
-  const optionFromChildren: Omit<IOption, 'type' | 'data' | 'width' | 'height'> = {};
+  const optionFromChildren: Omit<IOption, 'type' | 'data' | 'width' | 'height'> = {} as any;
 
   toArray(props.children).map((child, index) => {
     const parseOption = child && (child as any).type && (child as any).type.parseOption;
@@ -91,13 +98,17 @@ const parseOptionFromChildren = (props: Props) => {
           }
         : (child as any).props;
 
-      const optionResult = parseOption(childProps);
+      const optionResult = parseOption(childProps) as {
+        optionName: keyof Omit<IOption, 'type' | 'data' | 'width' | 'height'>;
+        isSingle: boolean;
+        option: any;
+      };
 
       if (optionResult.isSingle) {
-        optionFromChildren[optionResult.optionName] = optionResult.option;
+        optionFromChildren[optionResult.optionName] = optionResult.option as never;
       } else {
         if (!optionFromChildren[optionResult.optionName]) {
-          optionFromChildren[optionResult.optionName] = [];
+          optionFromChildren[optionResult.optionName] = [] as never;
         }
 
         optionFromChildren[optionResult.optionName].push(optionResult.option);
@@ -152,14 +163,15 @@ const BaseTable: React.FC<Props> = React.forwardRef((props, ref) => {
 
   const createTable = useCallback(
     (props: Props) => {
-      let vtable;
-      if (props.type === 'pivot-table') {
-        vtable = new VTable.PivotTable(props.container, parseOption(props));
-      } else if (props.type === 'pivot-chart') {
-        vtable = new VTable.PivotChart(props.container, parseOption(props));
-      } else {
-        vtable = new VTable.ListTable(props.container, parseOption(props));
-      }
+      // let vtable;
+      // if (props.type === 'pivot-table') {
+      //   vtable = new VTable.PivotTable(props.container, parseOption(props));
+      // } else if (props.type === 'pivot-chart') {
+      //   vtable = new VTable.PivotChart(props.container, parseOption(props));
+      // } else {
+      //   vtable = new VTable.ListTable(props.container, parseOption(props));
+      // }
+      const vtable = new props.vtableConstrouctor(props.container, parseOption(props));
       // vtable.scenegraph.stage.enableReactAttribute(ReactDOM);
       vtable.scenegraph.stage.reactAttribute = props.ReactDOM;
       vtable.scenegraph.stage.pluginService.register(new VTableReactAttributePlugin());
@@ -302,17 +314,25 @@ const BaseTable: React.FC<Props> = React.forwardRef((props, ref) => {
   );
 });
 
-export const createTable = <T extends Props>(componentName: string, type?: string, callback?: (props: T) => T) => {
+export const createTable = <T extends Props>(
+  componentName: string,
+  defaultProps?: Partial<T>,
+  callback?: (props: T) => T
+) => {
   const Com = withContainer<ContainerProps, T>(BaseTable as any, componentName, (props: T) => {
-    props.type = type;
+    // props.type = type;
 
-    if (callback) {
-      return callback(props);
+    if (defaultProps) {
+      return Object.assign(props, defaultProps);
     }
 
-    if (type) {
-      return { ...props, type };
-    }
+    // if (callback) {
+    //   return callback(props);
+    // }
+
+    // if (type) {
+    //   return { ...props, type };
+    // }
     return props;
   });
   Com.displayName = componentName;