+++ packages/vtable/examples/menu.ts
@@ -53,6 +53,10 @@ export const menus = [
       },
       {
         path: 'list',
+        name: 'list-transpose-frozen-row'
+      },
+      {
+        path: 'list',
         name: 'list-tree'
       },
       {
@@ -122,6 +126,14 @@ export const menus = [
       {
         path: 'list',
         name: 'list-large-columns'
+      },
+      {
+        path: 'list',
+        name: 'list-tree-group'
+      },
+      {
+        path: 'list',
+        name: 'list-group'
       }
     ]
   },