+++ packages/vtable/src/themes/theme.ts
@@ -74,6 +74,7 @@ export class TableTheme implements ITableThemeDefine {
   private _bottomFrozen: ITableThemeDefine['bottomFrozenStyle'] | null = null;
   private _rowHeader: ITableThemeDefine['rowHeaderStyle'] | null = null;
   private _body: ITableThemeDefine['bodyStyle'] | null = null;
+  private _groupTitle: ITableThemeDefine['groupTitleStyle'] | null = null;
   private _scroll: ITableThemeDefine['scrollStyle'] | null = null;
   private _tooltip: ITableThemeDefine['tooltipStyle'] | null = null;
   private _frameStyle: ITableThemeDefine['frameStyle'] | null = null;
@@ -407,6 +408,24 @@ export class TableTheme implements ITableThemeDefine {
     return this._body;
   }
 
+  get groupTitleStyle(): ITableThemeDefine['groupTitleStyle'] {
+    if (!this._groupTitle) {
+      const { obj, superTheme } = this.internalTheme;
+      if (!superTheme.groupTitleStyle && !obj.groupTitleStyle) {
+        return this._groupTitle;
+      }
+      const groupTitle: ThemeStyle = ingoreNoneValueMerge(
+        {},
+        this.defaultStyle,
+        superTheme.groupTitleStyle,
+        obj.groupTitleStyle
+      );
+      this._groupTitle = this.getStyle(groupTitle);
+    }
+
+    return this._groupTitle;
+  }
+
   get frameStyle(): ITableThemeDefine['frameStyle'] {
     if (!this._frameStyle) {
       const { obj, superTheme } = this.internalTheme;
@@ -436,7 +455,7 @@ export class TableTheme implements ITableThemeDefine {
         get shadowOffsetY(): number {
           return frameStyle.shadowOffsetY;
         },
-        get cornerRadius(): number {
+        get cornerRadius(): number | [number, number, number, number] {
           return frameStyle.cornerRadius;
         }
       };
@@ -875,7 +894,7 @@ export class TableTheme implements ITableThemeDefine {
       get cursor(): CursorPropertyDefine | undefined {
         return style.cursor;
       },
-      get textStick(): boolean | undefined {
+      get textStick(): boolean | 'vertical' | 'horizontal' | undefined {
         return style.textStick;
       },
       get marked(): MarkedPropertyDefine | undefined {