+++ packages/vtable/src/vrender.ts
@@ -54,7 +54,7 @@ export function registerForVrender() {
   registerRect();
   // registerRect3d();
   registerRichtext();
-  // registerShadowRoot();
+  registerShadowRoot();
   registerSymbol();
   registerText();
   // registerWrapText();
@@ -62,7 +62,7 @@ export function registerForVrender() {
   loadPoptip();
 }
 
-export { Direction } from '@visactor/vrender-core';
+export { Direction, Timeline } from '@visactor/vrender-core';
 export { GroupFadeIn } from '@visactor/vrender-core';
 export { GroupFadeOut } from '@visactor/vrender-core';